pyproject.toml
[tool.poetry]
name = "python-langchain"
version = "0.1.0"
description = ""
authors = ["work.ducmanh0323 <work.ducmanh0323@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.16"
langchain-experimental = "^0.0.57"
langchain-openai = "^0.1.3"
streamlit = "^1.33.0"
plotly = "^5.21.0"
load-dotenv = "^0.1.0"
faiss-cpu = "^1.8.0"
chromadb = "^0.4.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
